Add interchangeable heuristic executers.

+++ rust/src/constants.rs
@@ -1,2 +1,2 @@
 pub const MAX_TREE_SIZE: i32 = 40;
-pub const EDGE_COST: i32 = 1;
+pub const EDGE_COST: f32 = 1.0;
